// src/common/enums/auth/auth.enum.ts
export enum EError {
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  INVALID_EMAIL = 'INVALID_EMAIL',
  INVALID_PHONE_NUMBER = 'INVALID_PHONE_NUMBER',
  INVALID_AVATAR = 'INVALID_AVATAR',
  INVALID_IDENTITY_CARD = 'INVALID_IDENTITY_CARD',
  USER_EXIST = 'USER_EXIST',
  CANNOT_FIND_USER = 'CANNOT_FIND_USER',
  INVALID_USER_STATUS = 'INVALID_USER_STATUS',
  SIGN_UP_ERROR = 'SIGN_UP_ERROR',
  INVALID_INDETITY = 'INVALID_IDENTITY',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  TOKEN_IN_BLACKLIST = 'TOKEN_IN_BLACKLIST',
  DRIVER_NOT_FOUND = 'DRIVER_NOT_FOUND'
}

export enum EErrorDetail {
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  INVALID_EMAIL = 'INVALID_EMAIL',
  INVALID_PHONE_NUMBER = 'INVALID_PHONE_NUMBER',
  INVALID_AVATAR = 'INVALID_AVATAR',
  INVALID_IDENTITY_CARD = 'INVALID_IDENTITY_CARD',
  USER_EXIST = 'USER_EXIST',
  CANNOT_FIND_USER = 'CANNOT_FIND_USER',
  INVALID_USER_STATUS = 'INVALID_USER_STATUS',
  SIGN_UP_ERROR = 'Something went wrong, try again',
  INVALID_INDETITY = 'Try email or phone number',
  INVALID_REFRESH_TOKEN = 'Invalid refresh token',
  TOKEN_IN_BLACKLIST = 'Token in blacklist',
  DRIVER_NOT_FOUND = 'Driver not found'
}